Release of 1.9.5
========
Bugfixes
========
- Fixed a bug (redirection after recovering
was wrong)
- Fixed a bug which caused a wrong calculation of file
sizes
- Fixed a bug, the program did not checked correct if a file or
folder was existing
- Fixed a bug causing a wrong encoding of german
letters ä,ü,ö etc.*
- Validated the program by HTML5
- Fixed a bug that
it was not possible to change filenames
- Fixed a bug which allowed to
upload nothing to the file list
- Fixed a bug which allowed to upload a
.htaccess file to the file list
- Added a workaround to upload only
utf-8 encoded files.
- Fixed a bug that allowed to see the include
folder structure if a dialog failed
=================
Security
Bugfixes
=================
- Fixed a bug which allowed to upload a
.htaccess file to the file list (overwriting storage access
rule)
=================
Feature additions
=================
- Added new
setting Program_Mime_Use_DataBase to determine if the mime type will be
getted out of the database or the server's filesystem
- Added counter to
get the number of accesses on a share link
- Now the sharelink starts
with http:// or https:// instead of the server adress
- Added the usage
of the Program_Debug setting at some dialogs of the program
- Beta
feature*: Now it is possible to download directories as a zip file
- Now
it is possible to download the complete data
- Beta feature*: Folders
can now be shared (as zip file)
- Added tooltips for action icons
-
Added plugin interface
================
Features
removed
================
- Removed use of deprecated method in download
module

The rating of this build is still "unstable". Use with
caution!

*Known bug: The filenames are not converted properly
everywhere. It is possible that the filenames are broken, but the files
should be complete. It seems that here is an unresolved php bug
appearing on apache2.

+++ Includes/copy.inc.php
@@ -13,7 +13,11 @@
 				$file = mysqli_real_escape_string($connect,$_GET["file"]);	
 			else
 				$file = mysqli_real_escape_string($connect,$_POST["file"]);
-			copyFile($file,$dir);				
+			
+			if (fs_file_exists($file,$dir) == false){
+				copyFile($file,$dir);		
+			}
+			echo 	"existing:".fs_file_exists($file,$dir);		
 			mysqli_close($connect);	
 			$success = true;
 		}	
@@ -43,8 +47,11 @@
 	{		
 		echo "Command_Result:{$success}";
 		exit;		
-	}		
-	header("Location: ./index.php?module=list&dir=/");
-	exit;
-	
+	}
+	else{	
+		if ($GLOBALS["config"]["Program_Debug"] != 1){
+			header("Location: ./index.php?module=list&dir=".$_SESSION["currentdir"]);
+			exit;
+		}	
+	}
 ?>
